feat: [back] filter 매소드 부등호 변경 및 distance 인자 double형으로 바꿈 (#124)

+++ web-server/api/src/main/java/com/hotplace/api/controller/PlacePostController.java
@@ -39,7 +39,7 @@ public class PlacePostController {
     @GetMapping("/places/current")
     public ApiForm<List<PlaceResponse>> searchPlacesByCurrentLocation(@RequestParam ("latitude") Double latitude,
                                                                       @RequestParam ("longitude") Double longitude,
-                                                                      @RequestParam ("distance") Integer distance){
+                                                                      @RequestParam ("distance") Double distance){
         log.info("latitude: {}, longitude: {}", latitude, longitude);
         return placeService.searchPlacesByCurrentLocation(latitude, longitude, distance);
     }
@@ -48,7 +48,7 @@ public class PlacePostController {
     public ApiForm<List<PlaceResponse>> searchPlacesByLocation(
             @RequestParam ("latitude") Double latitude,
             @RequestParam ("longitude") Double longitude,
-            @RequestParam ("distance") Integer distance,
+            @RequestParam ("distance") Double distance,
             @RequestBody PlaceRequest requestDto
     ){
         log.info("latitude: {}, longitude: {}", latitude, longitude);
